userprofile and login.js handlers correction

+++ public/js/userprofile.js
@@ -6,7 +6,7 @@ const inputListName = document.getElementById('task-name');
 
 // Prevent going to another page
 const formEl = document.querySelector('.new-task-form');
-formEl.addEventListener("submit", (event)=>{
+formEl.addEventListener("submit", (event) => {
     event.preventDefault();
     saveList();
 })
@@ -18,38 +18,61 @@ const getList = (id) =>
         headers: {
             'Content-Type': 'application/json',
         },
-    });
+    }).then(response => response.json())
+        .then(data => {
+            console.log(data);
+            console.log(inputListNameText);
+
+            const tableBody = document.querySelector('#tbody');
+            const newRow = document.createElement('tr');
+            const listCell = document.createElement('td');
+
+            listCell.textContent = data;
+
+            newRow.appendChild(listCell);
+            tableBody.appendChild(newRow);
+        })
+
 
 const saveList = () => {
     const inputListNameText = inputListName.value;
-    console.log("Checking what we will send to express route")
-    console.log(inputListNameText)
 
     fetch('/api/lists', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
         },
-        body: JSON.stringify({list_name: inputListNameText}),
-    }).then(response=>response.json())
-    .then(data=>{
-        console.log(data);
-        // W3school
-        // var liEl = document.createElement("li")
-        // liEl.classList.add("some-class1")
-        // liEl.classList.add("some-class2")
-        // liEl.textContent = inputListName.value
-        // ulEl.appendChild(liEl)
-    })
+        body: JSON.stringify({ list_name: inputListNameText }),
+    }).then(response => response.json())
+        .then(data => {
+            console.log(data);
+            console.log(inputListNameText);
+
+            const tableBody = document.querySelector('#tbody');
+            const newRow = document.createElement('tr');
+            const listCell = document.createElement('td');
+
+            listCell.textContent = inputListNameText;
+            newRow.appendChild(listCell);
+            tableBody.appendChild(newRow);
+        })
 }
 
-const deleteList = (id) =>
-    fetch(`/api/lists/${id}`, {
+const deleteListBtn = async (e) => {
+    if (e.target.hasAttribute('data-id')) {
+      const id = e.target.getAttribute('data-id');
+  
+      const response = await fetch(`/api/projects/${id}`, {
         method: 'DELETE',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-    });
+      });
+      console.log(id)
+      if (response.ok) {
+        document.location.replace('/profile');
+      } else {
+        alert('Failed to delete project');
+      }
+    }
+  };
 
 const editList = (list) =>
     fetch(`/api/lists/${id}`, {
@@ -101,7 +124,7 @@ const handleListView = (e) => {
     e.preventDefault();
     activeNote = JSON.parse(e.target.parentElement.getAttribute('data-note'));
     renderActiveNote();
-  };
+};
